# openapi/openapi.yml
openapi: 3.0.0
servers:
  - description: Social Graph Development
    url: http://social-graph.localdev.me

info:
  version: "1.0.0"
  title: Go Social Graph API
  description: >-
    This is an example of using OAuth2 Access Code Flow in a specification to
    describe security to your API.

paths:
  /api/login:
    post:
      summary: Login and get a new token
      operationId: loginHandler
      tags:
        - auth
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                username:
                  type: string
                password:
                  type: string
                  format: password
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                properties:
                  token:
                    type: string

  /api/user:
    post:
      summary: Create a new user
      operationId: createUser
      tags:
        - user
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                username:
                  type: string
      responses:
        201:
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateUserResponse"

  /api/user/{id}/timeline:
    get:
      summary: Retrieve the timeline for a user
      operationId: timelineHandler
      tags:
        - user
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TimelineResponse"

  /api/user/{id}/follow/{from}:
    post:
      summary: Follows a user
      operationId: followHandler
      tags:
        - user
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
        - in: path
          name: from
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              properties:
                from:
                  type: string
                  format: uuid
      responses:
        200:
          description: OK

  /api/user/{id}/unfollow/{from}:
    post:
      summary: Unfollow a user
      operationId: unfollowHandler
      tags:
        - user
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
        - in: path
          name: from
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              properties:
                from:
                  type: string
                  format: uuid
      responses:
        200:
          description: OK

  /api/profile/{user_id}:
    get:
      summary: Retrieve user profile information
      operationId: profileHandler
      tags:
        - user
      parameters:
        - in: path
          name: user_id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProfileResponse"

  /api/post:
    post:
      summary: Create a new post for an user
      operationId: postHandler
      security:
        - BearerAuth: ['user:read']
      tags:
        - post
      requestBody:
        description: which user and content
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePostRequest"
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreatePostResponse"
  
  /api/post/{id}/repost:
    post:
      summary: Repost an user post
      operationId: repostHandler
      tags:
        - post
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        description: describes the repost
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RepostIntent'
      responses:
        200:
          description: OK

components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    CreateUserIntent:
      properties:
        username:
          type: string
    
    RepostIntent:
      properties:
        parent:
          type: string
          format: uuid
        quote:
          type: string
      required:
        - parent

    CreateUserResponse:
      properties:
        id:
          type: string
          format: uuid
        username:
          type: string
        created_at:
          type: string
          format: date-time
      required:
        - id
        - username
        - created_at

    CreatePostRequest:
      properties:
        user_id:
          type: string
          format: uuid
        content:
          type: string
          maxLength: 144
      required:
        - user_id
        - content

    CreatePostResponse:
      properties:
        id:
          type: string
          format: uuid
        content:
          type: string
      required:
        - id
        - content

    User:
      properties:
        id:
          type: string
          format: uuid
        username:
          type: string
        created_at:
          type: string
          format: date-time
      required:
        - id
        - username
        - created_at

    Post:
      properties:
        id:
          type: string
          format: uuid
        content:
          type: string
        parent:
          $ref: '#/components/schemas/UserPost'
        quote:
          type: string
        created_at:
          type: string
          format: date-time
      required:
        - id
        - content
        - created_at

    UserPost:
      properties:
        user:
          $ref: "#/components/schemas/User"
        post:
          $ref: "#/components/schemas/Post"
      required:
        - user
        - post

    TimelineResponse:
      properties:
        posts:
          type: array
          items:
            $ref: "#/components/schemas/UserPost"

    UserStats:
      properties:
        followers:
          type: integer
        following:
          type: integer
        posts_count:
          type: integer

    ProfileResponse:
      properties:
        user:
          $ref: "#/components/schemas/User"
        stats:
          $ref: "#/components/schemas/UserStats"
        posts:
          type: array
          items:
            $ref: "#/components/schemas/UserPost"